make Tract card class a static inner class of TractPage
This will allow us to better create other Card classes for other page types and in the Content model
